pyproject.toml
[tool.poetry]
name = "py-image-resize"
version = "0.1.0"
description = ""
authors = ["Juan Carlos Corona Romero <c@rlos.email>"]

[tool.poetry.dependencies]
python = "^3.10"
docopt-ng = "^0.8.1"
Pillow = "^9.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
